refactorization of hierarchical examples
