+++ examples/kinova_PUMA_3D_ModulationIK.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pybullet
-import time
 from pumafabrics.tamed_puma.create_environment.environments import trial_environments
 from pumafabrics.tamed_puma.examples_for_ROS.PUMA_3D_ModulationIK import PUMA_modulationIK
 """
@@ -21,10 +20,6 @@ class example_kinova_PUMA_modulationIK(PUMA_modulationIK):
         action = np.zeros(dof)
         self.initialize_environment()
         ob, *_ = self.env.step(action)
-        if self.params["nr_obst"] > 0:
-            self.obstacles = list(ob["robot_0"]["FullSensor"]["obstacles"].values())
-        else:
-            self.obstacles = []
         q_init = ob['robot_0']["joint_state"]["position"][0:dof]
 
         self.initialize_example(q_init=q_init)
@@ -40,15 +35,17 @@ class example_kinova_PUMA_modulationIK(PUMA_modulationIK):
             qdot = ob_robot["joint_state"]["velocity"][0:dof]
             if self.params["nr_obst"] > 0:
                 positions_obstacles = [ob_robot["FullSensor"]["obstacles"][self.params["nr_obst"]]["position"], ob_robot["FullSensor"]["obstacles"][self.params["nr_obst"]+1]["position"]]
+                obstacles = list(ob["robot_0"]["FullSensor"]["obstacles"].values())
             else:
                 positions_obstacles = []
+                obstacles = []
             goal_pos = [goal_pos[i] + self.params["goal_vel"][i] * self.params["dt"] for i in range(len(goal_pos))]
 
             runtime_arguments["q"] = q
             runtime_arguments["qdot"] = qdot
             runtime_arguments["positions_obstacles"] = positions_obstacles
             runtime_arguments["goal_pos"] = goal_pos
-            runtime_arguments["obstacles"] = self.obstacles
+            runtime_arguments["obstacles"] = obstacles
 
             action, self.GOAL_REACHED, dist_to_goal, self.IN_COLLISION, x_t_action = self.run(runtime_arguments)
             ob, *_ = self.env.step(action)
